requirements.txt
# FastAPI y servidor
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
jinja2>=3.1.6
python-multipart>=0.0.20

# Clientes asíncronos
openai>=1.6.0
anthropic>=0.8.0
google-generativeai>=0.3.0

# Concurrencia y utilidades
aiohttp>=3.9.0
httpx>=0.26.0

# Variables de entorno
python-dotenv>=1.0.0

# Métricas y análisis
numpy>=1.24.0
pandas>=2.1.0

# Logging y monitoreo
loguru>=0.7.0

# Validación y utilidades
typing-extensions>=4.8.0